feat(#6): keyboard navigation & a11y
