+++ src/index.css
@@ -47,22 +47,66 @@ body {
   border: dashed 1px blue;
 }
 
-#container > ul > li > button.pin {
-  position: absolute;
-  background: transparent;
-  border: none;
-  top: 5px;
-  margin: 0;
-  right: 10px;
-  padding: 5px;
-  width: 30px;
-  height: 30px;
-  cursor: pointer;
+#container > ul > li[data-pinned="true"] {
+  box-shadow: inset -5px 0px 0px 0px #8133ff;
 }
 
-#container > ul > li > button.pin > img {
+#container > ul > li:focus {
+  outline: none;
+}
+
+#container > ul > li:hover > div.entry-actions,
+#container > ul > li:focus-within > div.entry-actions,
+#container > ul > li:focus > div.entry-actions {
+  visibility: visible;
+}
+
+#container > ul > li > div.entry-actions {
+  display: flex;
+  visibility: hidden;
+  position: absolute;
+  background: #fff;
+  border: solid 1px rgba(0, 0, 0, 0.5);
+  border-radius: 10px;
+  width: 160px;
+  padding: 5px 10px;
+  justify-content: space-evenly;
+  top: -25px;
+  right: 10px;
+  transition: visibility 0.25s;
+}
+
+#container > ul > li > div.entry-actions > button {
   height: 100%;
   width: 100%;
+  margin: 0;
+  padding: 0;
+  cursor: pointer;
+  background: transparent;
+  border: none;
+  filter: saturate(0);
+  opacity: 0.5;
+  transition: filter 0.25s, opacity 0.25s;
+  outline: none;
+}
+
+#container > ul > li > div.entry-actions > button[data-active="true"],
+#container > ul > li > div.entry-actions > button:active,
+#container > ul > li > div.entry-actions > button:focus {
+  filter: saturate(1);
+  opacity: 1;
+}
+
+#container > ul > li > div.entry-actions > button[data-active="true"] {
+  filter: saturate(1);
+  opacity: 1;
+}
+
+#container > ul > li > div.entry-actions > button > img {
+  height: 20px;
+  width: 20px;
+  margin: 0;
+  padding: 0;
 }
 
 #container > ul > li:active {
