+++ src/index.html
@@ -15,5 +15,5 @@
 		<p>A tutorial is available in the menu bar.</p>
     <div id="container"></div>
   </body>
-  <script type="module" src="./render.js"></script>
+  <script type="module" src="../dist/render/main.js"></script>
 </html>
